Makes GetRandom throw an invalid_argument exception if the min paramenter is greater than the max paramenter.
Also put the GetRandom function into a header file
